[동호] 교수 과제(과제리스트조회,점수입력,전체파일다운,선택된파일다운)        security 사소한 오류 수정, log 개선

+++ choongangGroup/src/main/java/com/oracle/choongangGroup/sh/domain/Lecture.java
@@ -31,7 +31,7 @@ public class Lecture {
 	private String prof;
 
 	@Column(name = "lec_target_grade")
-	private int grade; //null 일때 Long으로 바꾸기
+	private Long grade; //null 일때 Long으로 바꾸기
 	
 	@Column(name = "lec_room")
 	private String room;
